rdrf #1546 made review changes
