+++ rdrf/rdrf/management/commands/pull_proms.py
@@ -1,7 +1,7 @@
 from django.core.management.base import BaseCommand
 from rdrf.models.definition.models import Registry
 
-from rdrf.services.rest.views.proms_api import PromsProcessor
+from rdrf.services.rest.views.proms_api import PromsSystemManager
 
 
 class Command(BaseCommand):
@@ -10,5 +10,5 @@ class Command(BaseCommand):
     def handle(self, *args, **options):
         for registry_model in Registry.objects.all():
             if registry_model.proms_system_url:
-                proms_processor = PromsProcessor(registry_model)
-                proms_processor.download_proms()
+                proms_manager = PromsSystemManager(registry_model)
+                proms_manager.download_proms()
